fix(clippy): is_none_or in sales allowlist gate; drop redundant f64 casts in VAT-rate test tables (pipeline clippy gate)

+++ services/mail-server/crates/billing-service/src/invoices.rs
@@ -1024,7 +1024,7 @@ mod tests {
         for (code, expected_rate) in &expected {
             let (rate, _) = calculate_vat(10000, code, None);
             assert_eq!(
-                rate, *expected_rate as f64,
+                rate, *expected_rate,
                 "Country {code} should have VAT rate {expected_rate}"
             );
         }
